requirements.txt
Flask==3.1.3
flask_google_cloud_logger==0.2.1
flask_talisman~=0.8.1
gunicorn~=25.1.0
firebase_admin~=5.0.3
